refactor dialog to re-use frequency components
